refactor: 멀티 모듈 구성 (api, domain)
