+++ api/build.gradle
@@ -1,0 +1,8 @@
+bootJar { enabled = true }
+jar { enabled = false }
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    implementation project(':domain')
+}
