+++ api/src/main/java/com/tago/api/ApiApplication.java
@@ -1,0 +1,13 @@
+package com.tago.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = {"com.tago.api", "com.tago.domain"})
+public class ApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ApiApplication.class, args);
+    }
+
+}
