+++ api/src/main/java/com/tago/api/auth/application/AuthController.java
@@ -1,0 +1,24 @@
+package com.tago.api.auth.application;
+
+
+import com.tago.domain.auth.domain.AuthTokens;
+import com.tago.api.auth.infra.kakao.KakaoLoginParams;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/auth")
+public class AuthController {
+    private final OAuthLoginService oAuthLoginService;
+
+    @PostMapping("/kakao")
+    public ResponseEntity<AuthTokens> loginKakao(@RequestBody KakaoLoginParams params) {
+        return ResponseEntity.ok(oAuthLoginService.login(params));
+    }
+
+}
