+++ api/src/main/java/com/tago/api/auth/application/OAuthLoginService.java
@@ -1,0 +1,41 @@
+package com.tago.api.auth.application;
+
+import com.tago.domain.auth.domain.oauth.OAuthInfoResponse;
+import com.tago.domain.auth.domain.oauth.OAuthLoginParams;
+import com.tago.domain.auth.domain.oauth.RequestOAuthInfoService;
+import com.tago.domain.auth.domain.AuthTokens;
+import com.tago.domain.auth.domain.AuthTokensGenerator;
+import com.tago.domain.domain.MemberRepository;
+import com.tago.domain.domain.Member;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class OAuthLoginService {
+    private final MemberRepository memberRepository;
+    private final AuthTokensGenerator authTokensGenerator;
+    private final RequestOAuthInfoService requestOAuthInfoService;
+
+    public AuthTokens login(OAuthLoginParams params) {
+        OAuthInfoResponse oAuthInfoResponse = requestOAuthInfoService.request(params);
+        Long memberId = findOrCreateMember(oAuthInfoResponse);
+        return authTokensGenerator.generate(memberId);
+    }
+
+    private Long findOrCreateMember(OAuthInfoResponse oAuthInfoResponse) {
+        return memberRepository.findByEmail(oAuthInfoResponse.getEmail())
+                .map(Member::getId)
+                .orElseGet(() -> newMember(oAuthInfoResponse));
+    }
+
+    private Long newMember(OAuthInfoResponse oAuthInfoResponse) {
+        Member member = Member.builder()
+                .email(oAuthInfoResponse.getEmail())
+                .nickname(oAuthInfoResponse.getNickname())
+                .oAuthProvider(oAuthInfoResponse.getOAuthProvider())
+                .build();
+
+        return memberRepository.save(member).getId();
+    }
+}
