+++ api/src/main/java/com/tago/api/auth/infra/kakao/KakaoLoginParams.java
@@ -1,0 +1,27 @@
+package com.tago.api.auth.infra.kakao;
+
+
+import com.tago.domain.auth.domain.oauth.OAuthLoginParams;
+import com.tago.domain.auth.domain.oauth.OAuthProvider;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+
+@Getter
+@NoArgsConstructor
+public class KakaoLoginParams implements OAuthLoginParams {
+    private String authorizationCode;
+
+    @Override
+    public OAuthProvider oAuthProvider() {
+        return OAuthProvider.KAKAO;
+    }
+
+    @Override
+    public MultiValueMap<String, String> makeBody() {
+        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
+        body.add("code", authorizationCode);
+        return body;
+    }
+}
