+++ domain/src/main/java/com/tago/domain/DomainApplication.java
@@ -1,0 +1,13 @@
+package com.tago.domain;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DomainApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DomainApplication.class, args);
+    }
+
+}
