+++ domain/src/main/java/com/tago/domain/auth/domain/oauth/OAuthApiClient.java
@@ -1,0 +1,7 @@
+package com.tago.domain.auth.domain.oauth;
+
+public interface OAuthApiClient {
+    OAuthProvider oAuthProvider();
+    String requestAccessToken(OAuthLoginParams params);
+    OAuthInfoResponse requestOauthInfo(String accessToken);
+}
