+++ domain/src/main/java/com/tago/domain/auth/domain/oauth/OAuthInfoResponse.java
@@ -1,0 +1,7 @@
+package com.tago.domain.auth.domain.oauth;
+
+public interface OAuthInfoResponse {
+    String getEmail();
+    String getNickname();
+    OAuthProvider getOAuthProvider();
+}
