+++ domain/src/main/java/com/tago/domain/auth/domain/oauth/OAuthLoginParams.java
@@ -1,0 +1,9 @@
+package com.tago.domain.auth.domain.oauth;
+
+import org.springframework.util.MultiValueMap;
+
+//OAuth요청을 위한 파라미터 값들을 갖고 있는 인터페이스
+public interface OAuthLoginParams {
+    OAuthProvider oAuthProvider();
+    MultiValueMap<String, String> makeBody();
+}
