+++ domain/src/main/java/com/tago/domain/auth/domain/oauth/OAuthProvider.java
@@ -1,0 +1,6 @@
+package com.tago.domain.auth.domain.oauth;
+
+public enum OAuthProvider {
+
+    KAKAO
+}
