+++ domain/src/main/java/com/tago/domain/domain/Member.java
@@ -1,0 +1,35 @@
+package com.tago.domain.domain;
+
+
+import com.tago.domain.auth.domain.oauth.OAuthProvider;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String email;
+
+    private String nickname;
+
+    private OAuthProvider oAuthProvider;
+
+    @Builder
+    public Member(String email, String nickname, OAuthProvider oAuthProvider){
+        this.email = email;
+        this.nickname = nickname;
+        this.oAuthProvider = oAuthProvider;
+    }
+
+}
