+++ domain/src/main/java/com/tago/domain/domain/MemberRepository.java
@@ -1,0 +1,10 @@
+package com.tago.domain.domain;
+
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String email);
+}
