+++ domain/src/test/java/com/tago/domain/DomainApplicationTests.java
@@ -1,0 +1,13 @@
+package com.tago.domain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DomainApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
