--- src/main/java/com/api/tago/TagoApplication.java
@@ -1,13 +1,0 @@
-package com.api.tago;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TagoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TagoApplication.class, args);
-	}
-
-}
