--- src/main/java/com/api/tago/auth/application/AuthController.java
@@ -1,24 +1,0 @@
-package com.api.tago.auth.application;
-
-
-import com.api.tago.auth.domain.AuthTokens;
-import com.api.tago.auth.infra.kakao.KakaoLoginParams;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/auth")
-public class AuthController {
-    private final OAuthLoginService oAuthLoginService;
-
-    @PostMapping("/kakao")
-    public ResponseEntity<AuthTokens> loginKakao(@RequestBody KakaoLoginParams params) {
-        return ResponseEntity.ok(oAuthLoginService.login(params));
-    }
-
-}
