--- src/main/java/com/api/tago/auth/domain/AuthTokens.java
@@ -1,21 +1,0 @@
-package com.api.tago.auth.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthTokens {
-    private String accessToken;
-    private String refreshToken;
-    private String grantType;
-    private Long expiresIn;
-
-    public static AuthTokens of(String accessToken, String refreshToken, String grantType, Long expiresIn) {
-        return new AuthTokens(accessToken, refreshToken, grantType, expiresIn);
-    }
-}
