--- src/main/java/com/api/tago/auth/domain/oauth/OAuthApiClient.java
@@ -1,7 +1,0 @@
-package com.api.tago.auth.domain.oauth;
-
-public interface OAuthApiClient {
-    OAuthProvider oAuthProvider();
-    String requestAccessToken(OAuthLoginParams params);
-    OAuthInfoResponse requestOauthInfo(String accessToken);
-}
