--- src/main/java/com/api/tago/auth/domain/oauth/OAuthInfoResponse.java
@@ -1,7 +1,0 @@
-package com.api.tago.auth.domain.oauth;
-
-public interface OAuthInfoResponse {
-    String getEmail();
-    String getNickname();
-    OAuthProvider getOAuthProvider();
-}
