--- src/main/java/com/api/tago/auth/domain/oauth/OAuthProvider.java
@@ -1,6 +1,0 @@
-package com.api.tago.auth.domain.oauth;
-
-public enum OAuthProvider {
-
-    KAKAO
-}
