--- src/main/java/com/api/tago/auth/infra/kakao/KakaoInfoResponse.java
@@ -1,43 +1,0 @@
-package com.api.tago.auth.infra.kakao;
-
-import com.api.tago.auth.domain.oauth.OAuthInfoResponse;
-import com.api.tago.auth.domain.oauth.OAuthProvider;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-
-@Getter
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class KakaoInfoResponse implements OAuthInfoResponse {
-
-    @JsonProperty("kakao_account")
-    private KakaoAccount kakaoAccount;
-
-    @Getter
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    static class KakaoAccount {
-        private KakaoProfile profile;
-        private String email;
-    }
-
-    @Getter
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    static class KakaoProfile {
-        private String nickname;
-    }
-
-    @Override
-    public String getEmail() {
-        return kakaoAccount.email;
-    }
-
-    @Override
-    public String getNickname() {
-        return kakaoAccount.profile.nickname;
-    }
-
-    @Override
-    public OAuthProvider getOAuthProvider() {
-        return OAuthProvider.KAKAO;
-    }
-}
