--- src/main/java/com/api/tago/auth/infra/kakao/KakaoLoginParams.java
@@ -1,27 +1,0 @@
-package com.api.tago.auth.infra.kakao;
-
-
-import com.api.tago.auth.domain.oauth.OAuthLoginParams;
-import com.api.tago.auth.domain.oauth.OAuthProvider;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-
-@Getter
-@NoArgsConstructor
-public class KakaoLoginParams implements OAuthLoginParams {
-    private String authorizationCode;
-
-    @Override
-    public OAuthProvider oAuthProvider() {
-        return OAuthProvider.KAKAO;
-    }
-
-    @Override
-    public MultiValueMap<String, String> makeBody() {
-        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("code", authorizationCode);
-        return body;
-    }
-}
