--- src/main/java/com/api/tago/member/domain/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.api.tago.member.domain;
-
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-}
