--- src/main/resources/application.yml
@@ -1,24 +1,0 @@
-spring:
-  jpa:
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-    defer-datasource-initialization: true
-
-  datasource:
-    url: jdbc:h2:mem:testdb
-
-  h2:
-    console:
-      enabled: true
-
-jwt:
-  secret-key: Z29nby10bS1zZXJ2ZXItZGxyamVvYW9yb3JodG9kZ290c3Atam9vbmdhbmduaW0teWVvbHNpbWhpaGFsZ2VveW8K
-
-oauth:
-  kakao:
-    client-id: bf9b0d47a797d8efdd9ba087afaf9d5f
-    url:
-      auth: https://kauth.kakao.com
-      api: https://kapi.kakao.com
