--- src/test/java/com/api/tago/TagoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.api.tago;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TagoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
